make names match convention of the rest of the project
finish implementing thumbnail caching.
