+++ .gitignore
@@ -60,6 +60,8 @@ UserSettings/EditorUserSettings.asset
 
 Packages/packages-lock.json
 
+material_browser_cache.json
+
 # temporary (remove me before merging material browser PR)
-/Assets/Real Materials/
-/Assets/Real Materials.meta
+/Assets/Surfaces Beta 0.4/
+/Assets/Surfaces Beta 0.4.meta
