+++ Packages/com.chisel.editor/Chisel/Editor/Editor/MaterialBrowser/ChiselEmbeddedTextures.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 9dac7a578e4a432797e3a3b0cde953ed
+timeCreated: 1601856748
