+++ Packages/com.chisel.editor/Chisel/Editor/Editor/MaterialBrowser/ChiselMaterialBrowserUtilities.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: d081047e998f4e5396bd178535186bd1
+timeCreated: 1601856923
