+++ Packages/com.chisel.editor/Chisel/Editor/Editor/MaterialBrowser/ChiselMaterialThumbnailRenderer.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: f96e7427018a4141911591d290a1506c
+timeCreated: 1601856748
