--- Packages/com.chisel.editor/Editor Resources/MaterialBrowserCache.asset
@@ -1,14 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 566b787ede4f4f8fa61fe7ca97f6c911, type: 3}
-  m_Name: MaterialBrowserCache
-  m_EditorClassIdentifier:
