--- Packages/com.chisel.editor/Editor Resources/MaterialBrowserCache.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8a16b375fadd3314aaf8e39b9044f8d7
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
