+++ ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2020.2.0b4
-m_EditorVersionWithRevision: 2020.2.0b4 (7b778e331e0b)
+m_EditorVersion: 2020.2.0b5
+m_EditorVersionWithRevision: 2020.2.0b5 (e2067236bd5c)
