+++ README.md
@@ -10,7 +10,7 @@ Chisel is an extension for **Unity** that allows for _rapid_, iterative, **non d
 * If you have questions be sure to ask in the discord! **[Click here for to join the official Discord server!](https://discord.gg/zttNkPQ)**
 
 Notes:
-- Chisel requires Unity 2020.2.b04 or newer
+- Chisel requires Unity 2020.2.b05 or newer
 - UX is still under development
 - CSG algorithm is still under development
 - It is *not yet* ready for production, but please feel free to try it out, or better yet, help out!
